% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QI_mergePos.R
\name{QI_mergePos}
\alias{QI_mergePos}
\title{Merge Positive and Negative Metabolomics Data}
\usage{
QI_mergePos(posData, negData)
}
\arguments{
\item{posData}{A data frame containing the positive ion mode metabolomics data.}

\item{negData}{A data frame containing the negative ion mode metabolomics data.}
}
\value{
A data frame with merged data from positive and negative ion mode, retaining entries
        with the highest score and, in case of duplicates, the highest Fragmentation Score.
}
\description{
This function merges positive and negative metabolomics data based on the Compound ID.
It retains the entry with the highest score for each Compound ID, and if there are duplicates,
it then selects the entry with the highest Fragmentation Score.
}
\examples{
# Assuming posData and negData are data frames with metabolomics data

}
\seealso{
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{group_by}}
}
\author{
Chuanping Zhao
}
\keyword{data}
\keyword{merging}
\keyword{metabolomics}
