% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QI_mergeQQ.R
\name{QI_mergeQQ}
\alias{QI_mergeQQ}
\title{Merge Qualitative and Quantitative Metabolomics Data exported from Progenesis QI software}
\usage{
QI_mergeQQ(
  identical_path = "inputfile/pos_identification.csv",
  intensity_path = "inputfile/pos_measurement.csv",
  score_value = 0,
  polarity = "pos"
)
}
\arguments{
\item{identical_path}{The file path to the qualitative data exported from Progenesis QI.
This is a required parameter. Supported formats are CSV and XLSX, and TXT.}

\item{intensity_path}{The file path to the quantitative data exported from Progenesis QI.
This is a required parameter. Supported formats are CSV, XLSX, and TXT.}

\item{score_value}{The score threshold for filtering MS1 levels.
This is a required parameter. The default value is 0.}

\item{polarity}{The polarity mode of the mass spectrometry data.
This is a required parameter. The default value is "pos".}
}
\value{
A data frame containing the merged qualitative and quantitative data, filtered by score, with
duplicates removed, and augmented with polarity information.
}
\description{
This function is specifically designed to merge qualitative and quantitative metabolomics data
exported from Progenesis QI software. It processes and merges datasets in CSV, XLSX, and TXT formats,
prioritizing quantitative data when overlaps occur. The function filters data by score, handles
duplicate entries, and augments the data with polarity information, making it ideal for Progenesis QI
analysis workflows.
}
\seealso{
\code{\link[dplyr]{left_join}}, \code{\link[readr]{read_csv}}
}
\author{
Chuanping Zhao
}
\keyword{data}
\keyword{merging}
