% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{check_and_install_packages}
\alias{check_and_install_packages}
\title{Read Data from Various File Types}
\usage{
check_and_install_packages(packages)
}
\arguments{
\item{file_path}{The file path to the data file to be read.}
}
\value{
Returns a data frame containing the data from the file.
}
\description{
This function reads data from files with different extensions (e.g., CSV, XLSX, TXT) and returns
the data as a data frame or data table. It automatically installs required packages if they are
not already installed. For CSV files, it uses the base R `read.csv` function.
}
\note{
If a required package is not installed, the function will attempt to install it automatically.
}
\seealso{
\code{\link[base]{read.csv}}, \code{\link[readxl]{read_excel}}, \code{\link[data.table]{fread}}
}
\author{
Chuanping Zhao
}
\keyword{file}
\keyword{io}
